requirements.txt
numpy
matplotlib
pyvisa
pyserialnumpy
matplotlib
pyvisa
pyserial